升级Springboot 2.5.7、redisson 3.16.4、mybatis-plus 3.4.3.4、pagehelper 1.4.0、guava 31.0.1-jre、tika-core 2.1.0； mybatis-plus新版本selectCount返回值修改为Long的处理；

+++ foodie-shop-pojo/src/main/java/cn/myst/web/pojo/vo/CommentLevelCountsVO.java
@@ -16,16 +16,16 @@ import java.io.Serializable;
 public class CommentLevelCountsVO implements Serializable {
 
     @ApiModelProperty(value = "所有评价的数量")
-    private Integer totalCounts;
+    private Long totalCounts;
 
     @ApiModelProperty(value = "好评的数量")
-    private Integer goodCounts;
+    private Long goodCounts;
 
     @ApiModelProperty(value = "中评的数量")
-    private Integer normalCounts;
+    private Long normalCounts;
 
     @ApiModelProperty(value = "差评的数量")
-    private Integer badCounts;
+    private Long badCounts;
 
     private static final long serialVersionUID = -3616983867576705847L;
 }
